SBR DSP: use intptr_t for the ixh parameter.

+++ libavcodec/arm/sbrdsp_init_arm.c
@@ -30,7 +30,7 @@ void ff_sbr_qmf_post_shuffle_neon(float W[32][2], const float *z);
 void ff_sbr_qmf_deint_neg_neon(float *v, const float *src);
 void ff_sbr_qmf_deint_bfly_neon(float *v, const float *src0, const float *src1);
 void ff_sbr_hf_g_filt_neon(float (*Y)[2], const float (*X_high)[40][2],
-                           const float *g_filt, int m_max, int ixh);
+                           const float *g_filt, int m_max, intptr_t ixh);
 void ff_sbr_hf_gen_neon(float (*X_high)[2], const float (*X_low)[2],
                         const float alpha0[2], const float alpha1[2],
                         float bw, int start, int end);
